+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 4. Add test to see if import_drg_route_distribution_id exists during Terraform generation.
 5. Fix issue where network entity in route rule was not filtered by vcn id
 6. Resolve missing Icon in Relationship view
+7. Update Price Calculator to check if resource in okit model before attempting to access it (Issue: #519)
 
 
 ## Version 0.34.0